requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
gigachat>=0.1.2
jinja2==3.1.5
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
python-dotenv==1.0.1
PyMuPDF==1.24.14
Pillow==11.0.0
